Rename filnet directory to net and move pubsub and the plumbing network object under it. (#2190)
