+++ net/bootstrap_test.go
@@ -1,4 +1,4 @@
-package filnet
+package net
 
 import (
 	"context"
