+++ net/network.go
@@ -1,12 +1,11 @@
-package ntwk
+package net
 
 import (
 	"gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
 	"gx/ipfs/QmZZseAa9xcK6tT3YpaShNUAEpyRAoWmUL5ojH3uGNepAc/go-libp2p-metrics"
 	"gx/ipfs/Qmd52WKRSwrBK5gUaJKawryZQ5by6UbNB8KVW2Zy6JtbyW/go-libp2p-host"
 
-	"github.com/filecoin-project/go-filecoin/filnet"
-	"github.com/filecoin-project/go-filecoin/pubsub"
+	"github.com/filecoin-project/go-filecoin/net/pubsub"
 )
 
 // Network is a unified interface for dealing with libp2p
@@ -15,7 +14,7 @@ type Network struct {
 	*pubsub.Subscriber
 	*pubsub.Publisher
 	metrics.Reporter
-	*filnet.Router
+	*Router
 }
 
 // New returns a new Network
@@ -23,7 +22,7 @@ func New(
 	host host.Host,
 	publisher *pubsub.Publisher,
 	subscriber *pubsub.Subscriber,
-	router *filnet.Router,
+	router *Router,
 	reporter metrics.Reporter,
 ) *Network {
 	return &Network{
