+++ node/message.go
@@ -3,7 +3,7 @@ package node
 import (
 	"context"
 
-	"github.com/filecoin-project/go-filecoin/pubsub"
+	"github.com/filecoin-project/go-filecoin/net/pubsub"
 	"github.com/filecoin-project/go-filecoin/types"
 )
